pyproject.toml
[project]
name = "reddit-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asyncpraw>=7.8.1",
    "google-genai>=1.16.1",
    "mcp[cli]>=1.9.1",
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.1.0",
]
